Add more fields to error
