+++ lib/props/index.js
@@ -1,18 +1,35 @@
 const FIELDS = [
+
+	// Error native getters
 	'message',
 	'stack',
+	'code',
 
 	// inherited (not own property)
 	'name',
-
-	// optional
-	'code',
-	'details',
 
 	// non standard browser fields
 	'fileName',
 	'lineNumber',
 	'columnNumber',
+
+	// SystemError
+	'address',
+	'dest',
+	'errno',
+	'info',
+	'path',
+	'port',
+	'syscall',
+
+	// OpenSSL error properties
+	'opensslErrorStack',
+	'function',
+	'library',
+	'reason',
+
+	// custom
+	'details',
 ];
 
 /**
